move admin page part to other route
